+++ src/ChronicleServiceProvider.php
@@ -6,6 +6,8 @@ use Chronicle\Console\Commands\CreateCheckpointCommand;
 use Chronicle\Console\Commands\ExportCommand;
 use Chronicle\Console\Commands\InstallCommand;
 use Chronicle\Console\Commands\ReportCommand;
+use Chronicle\Console\Commands\ShowEntryCommand;
+use Chronicle\Console\Commands\StatsCommand;
 use Chronicle\Console\Commands\VerifyEntryCommand;
 use Chronicle\Console\Commands\VerifyExportCommand;
 use Chronicle\Context\QueueJobContext;
@@ -78,6 +80,8 @@ class ChronicleServiceProvider extends ServiceProvider
                 VerifyExportCommand::class,
                 InstallCommand::class,
                 ReportCommand::class,
+                StatsCommand::class,
+                ShowEntryCommand::class,
             ]);
         }
     }